requirements.txt
openai>=0.28
Pillow>=9.0
requests>=2.28
python-dotenv>=0.21  # Optional but recommended